Breakout player code into its own file.
Also add one function from player into socket to prevent circular
dependencies.
