+++ lib/index.js
@@ -1,3 +1,4 @@
 module.exports = function() {
+	this.Player = require('./player.js');
 	this.Socket = require('./socket.js');
 }
