+++ lib/socket.js
@@ -1,20 +1,20 @@
-module.exports = {
+var Socket = {
 	// connect new socket
 	connect: function(socket){
 		debug_message('Connecting socket');
 		socket.start_time = now();
 		socket.initials = 'unk';
 		// reconnect(socket); // will create player if it doesn't work.
-		create_player(socket);
-		set_update_player_interval(socket);
+		Player.create_player(socket);
+		Socket.set_update_player_interval(socket);
 
 		socket.on('set_initials', function(data){
 			if (data && data != socket.initials){
 				socket.initials = data.trim().toUpperCase().substring(0, 3);
 				
 				debug_message('Incoming initials: '+socket.initials);
-				update_initials(socket);
-				reset_score(socket);
+				Player.update_initials(socket);
+				Player.reset_score(socket);
 			}
 		});
 
@@ -60,7 +60,7 @@ module.exports = {
 		});
 
 		socket.on('disconnect', function(){
-			disconnect(socket);
+			Socket.disconnect(socket);
 		});
 	},
 
@@ -78,5 +78,22 @@ module.exports = {
 			score = elapsed_time / 60.0;
 
 		socket.emit('score_minutes', score.toFixed(3));
+	},
+
+	set_update_player_interval: function(socket) {
+		if (socket.update_player_interval){
+			clearInterval(socket.update_player_interval);
+		}
+
+		debug_message('Setting update player interval');
+
+		socket.update_player_interval = setInterval(function(){
+			Player.ping_player(socket);
+			Socket.send_score(socket);
+			send_team_online(socket);
+		}, 5000);
+
 	}
 };
+
+module.exports = Socket;
