+++ server.js
@@ -35,31 +35,6 @@ function db_call(query, params, error_msg, success_callback){
 	});
 }
 
-// Broadcast
-// broadcast who's online boards
-function emit_online_players(){
-	db_call("SELECT `initials`, `team`, TIMESTAMPDIFF(SECOND, connected_at, last_ping) AS `score_seconds`" +
-		" FROM `players`" +
-		" WHERE TIMESTAMPDIFF(MINUTE, last_ping, NOW()) < 1" +
-		" ORDER BY `score_seconds` DESC",
-		[],
-		"Querying for online players",
-		function(results){
-			io.sockets.emit('online', results);
-			debug_message('Online Player board updated.');
-		});
-}
-
-// broadcast high score boards for players
-function emit_high_scores(){
-	db_call("SELECT `initials`, TIMESTAMPDIFF(SECOND, connected_at, last_ping) AS `score_seconds`" +
-		" FROM `players`" +
-		" ORDER BY `score_seconds` DESC LIMIT 10", [], "Querying for high scores", function(results){
-			io.sockets.emit('high_scores', results);
-			debug_message('High Scores board updated.');
-		});
-}
-
 // broadcast board for high team scores
 function emit_team_high_scores(){
 	db_call(" SELECT" +
@@ -76,26 +51,6 @@ function emit_team_high_scores(){
 			" `score_seconds` DESC LIMIT 5;", [], "Querying for team high scores", function(results){
 			io.sockets.emit('team_high_scores', results);
 			debug_message('Team High Scores board updated.');
-		});
-}
-
-// Updates for player
-// set 0 score (eg for new initials)
-function reset_score(socket){
-	if (!socket || !socket.player_id){
-		log_message('Error resetting score: player_id is falsey.');
-		return;
-	}
-
-	db_call("UPDATE `players` SET `connected_at` = CURRENT_TIMESTAMP, `last_ping` = CURRENT_TIMESTAMP WHERE `id` = ?",
-		[socket.player_id],
-		"Resetting score",
-		function(results){
-			if (results.changedRows > 0){
-				debug_message("player score updated.");
-				socket.start_time = now();
-				socket.emit('score', '0.000');
-			}
 		});
 }
 
@@ -118,97 +73,6 @@ function update_team(socket){
 		});
 }
 
-// save player initials
-function update_initials(socket){
-	if (!socket || !socket.player_id || !socket.initials){
-		log_message('Error updating initials: socket.initials is falsey.');
-		return;
-	}
-
-	db_call("UPDATE `players` SET `initials` = ? WHERE `id` = ?",
-		[socket.initials, socket.player_id],
-		"Updating player initials",
-		function(results){
-			if (results.changedRows === 0){
-				debug_message("Warning updating player initials: No changed rows.");
-			} else {
-				debug_message("Player initials updated: " +socket.initials);
-			}
-		});
-}
-
-// create player
-function create_player(socket){
-	var msg;
-
-	if (!socket){
-		msg = 'Cannot create player: socket is falsey.';
-		log_message(msg);
-		return;
-	}
-
-	if (!socket || !socket.initials){
-		msg = 'Cannot create player: socket.initials is falsey.';
-		log_message(msg);
-		socket.emit('error', msg);
-		socket.disconnect();
-
-		return;
-	}
-
-	var initials = socket.initials,
-		ip = socket.request.connection.remoteAddress,
-		team = null;
-
-	if (socket.team){
-		team = socket.team;
-	}
-
-	pool.getConnection(function(error, connection){
-		if (error){
-			msg = 'Cannot create player: Cannot get connection';
-			log_message(msg+': '+error);
-			socket.emit('error', msg);
-			socket.disconnect();
-			return;
-		}
-
-		var query = connection.query("INSERT INTO `players` (`ip`, `initials`, `team`) VALUES (INET_ATON(?), ?, ?)", [ip, initials, team], function(error, result){
-			connection.release();
-
-			if (error){
-				log_message("ERROR creating player: "+error);
-				socket.emit('error', 'ERROR creating player: Query error.');
-				socket.disconnect();
-				return;
-			}
-
-			socket.player_id = result.insertId;
-			debug_message('player '+initials+' from '+ip+' inserted (id='+result.insertId+')');
-		});
-	});
-}
-
-// ping player: updates players most recent ping time in table to update who's online and update their scores too.
-function ping_player(socket){
-	if (!socket){
-		log_message('Error updating player ping time: socket is falsey.');
-		return;
-	} else if (!socket.player_id){
-		log_message('Error updating player ping time: socket.player_id is falsey.');
-		return;
-	}
-
-	db_call("UPDATE `players` SET `last_ping` = CURRENT_TIMESTAMP WHERE `id` = ?",
-		[socket.player_id],
-		"Updating player last_ping",
-		function(results){
-			if (results.changedRows > 0){
-				debug_message("Player last ping updated.");
-			}
-		});
-}
-
 function send_team_online(socket){
 	if (!socket || !socket.team){
 		return;
@@ -221,20 +85,6 @@ function send_team_online(socket){
 		"Getting team for player "+socket.initials, function(results){
 			socket.emit('team_online', results);
 		});
-}
-
-function set_update_player_interval(socket){
-	if (socket.update_player_interval){
-		clearInterval(socket.update_player_interval);
-	}
-
-	debug_message('Setting update player interval');
-
-	socket.update_player_interval = setInterval(function(){
-		ping_player(socket);
-		send_score(socket);
-		send_team_online(socket);
-	}, 5000);
 }
 
 
@@ -294,8 +144,8 @@ _prompt.get({properties: {
 	io.sockets.on('connection', Socket.connect);
 
 	setInterval(function(){
-		emit_online_players();
-		emit_high_scores();
+		Player.emit_online_players();
+		Player.emit_high_scores();
 		emit_team_high_scores();
 	}, 5000);
